server/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2
pydantic
python-dotenv
starlette
